requirements.txt
fastapi
uvicorn
google-search-results
youtube_transcript_api
gtts
transformers
torch
python-dotenv